+++ gulpfile.js
@@ -2,111 +2,8 @@ var gulp = require("gulp");
 var uglify = require("gulp-uglify");
 var concat = require("gulp-concat");
 
-var fileSets = [
-	"./js/init.js",
-	"./js/canvas/init.js",
-	"./js/draw/init.js",
-	"./js/physics/init.js",
-	"./js/canvas/animation/main.js",
-	"./js/main/importka.js",
-	"./js/main/loop.js",
-	"./js/ui/init.js",
-	"./js/canvas/animation/core.js",
-	"./js/canvas/animation/animset.js",
-	"./js/canvas/utils/configure.js",
-	"./js/canvas/utils/deconfigure.js",
-	"./js/canvas/utils/resize.js",
-	"./js/canvas/utils/update.js",
-	"./js/canvas/utils/coordinates.js",
-	"./js/physics/minilibs/equation.js",
-	"./js/physics/minilibs/vector2/init.js",
-	"./js/physics/minilibs/collisiondetection.js",
-	"./js/ui/basic.js",
-	"./js/ui/updateset.js",
-	"./js/draw/colors/fill.js",
-	"./js/draw/colors/stroke.js",
-	"./js/draw/colors/nostroke.js",
-	"./js/draw/colors/nofill.js",
-	"./js/draw/colors/getfill.js",
-	"./js/draw/colors/getstroke.js",
-	"./js/draw/settings/strokeweight.js",
-	"./js/draw/settings/getstrokeweight.js",
-	"./js/draw/settings/rectmode.js",
-	"./js/draw/settings/ellipsemode.js",
-	"./js/draw/settings/imagemode.js",
-	"./js/draw/settings/getrectmode.js",
-	"./js/draw/settings/getellipsemode.js",
-	"./js/draw/settings/getimagemode.js",
-	"./js/draw/settings/strokejoin.js",
-	"./js/draw/settings/strokecap.js",
-	"./js/draw/settings/getstrokejoin.js",
-	"./js/draw/settings/getstrokecap.js",
-	"./js/draw/settings/textfont.js",
-	"./js/draw/settings/gettextfont.js",
-	"./js/draw/settings/textsize.js",
-	"./js/draw/settings/gettextsize.js",
-	"./js/draw/settings/gettextalign.js",
-	"./js/draw/settings/gettextbaseline.js",
-	"./js/draw/settings/textalign.js",
-	"./js/draw/settings/loadfont.js",
-	"./js/draw/settings/loadfontset.js",
-	"./js/draw/settings/textlinespacing.js",
-	"./js/draw/settings/gettextlinespacing.js",
-	"./js/draw/transform/popmatrix.js",
-	"./js/draw/transform/pushmatrix.js",
-	"./js/draw/transform/resetmatrix.js",
-	"./js/draw/transform/rotate.js",
-	"./js/draw/transform/scale.js",
-	"./js/draw/transform/skew.js",
-	"./js/draw/transform/translate.js",
-	"./js/draw/shapes/arc.js",
-	"./js/draw/shapes/background.js",
-	"./js/draw/shapes/beginshape.js",
-	"./js/draw/shapes/beziervertex.js",
-	"./js/draw/shapes/ellipse.js",
-	"./js/draw/shapes/endshape.js",
-	"./js/draw/shapes/line.js",
-	"./js/draw/shapes/point.js",
-	"./js/draw/shapes/poly.js",
-	"./js/draw/shapes/quadvertex.js",
-	"./js/draw/shapes/rect.js",
-	"./js/draw/shapes/text.js",
-	"./js/draw/shapes/vertex.js",
-	"./js/draw/image/images.js",
-	"./js/draw/image/loadimage.js",
-	"./js/draw/image/loadimageset.js",
-	"./js/draw/image/image.js",
-	"./js/draw/image/get.js",
-	"./js/physics/minilibs/vector2/arithmetic.js",
-	"./js/physics/minilibs/vector2/convert.js",
-	"./js/physics/minilibs/vector2/direction.js",
-	"./js/physics/minilibs/vector2/dot.js",
-	"./js/physics/minilibs/vector2/magnitude.js",
-	"./js/physics/minilibs/vector2/reflect.js",
-	"./js/physics/minilibs/vector2/canvasmap.js",
-	"./js/ui/rectbasic.js",
-	"./js/ui/ellipsebasic.js",
-	"./js/physics/entities/circle/init.js",
-	"./js/physics/entities/line.js",
-	"./js/physics/entities/constraints/constraint.js",
-	"./js/ui/rectbutton.js",
-	"./js/ui/ellipsebutton.js",
-	"./js/physics/entities/constraints/bungee.js",
-	"./js/physics/entities/constraints/compressor.js",
-	"./js/physics/entities/constraints/cord.js",
-	"./js/physics/entities/constraints/extender.js",
-	"./js/physics/entities/constraints/rod.js",
-	"./js/physics/entities/constraints/spring.js",
-	"./js/physics/entities/circle/core.js",
-	"./js/physics/entities/circle/force.js",
-	"./js/physics/entities/circle/collideline.js",
-	"./js/physics/entities/circle/collidecircle.js",
-	"./js/ui/rectdragger.js",
-	"./js/ui/ellipsedragger.js"
-];
-
 gulp.task("build", function() {
-	gulp.src(fileSets).pipe(concat("kaphy.js")).pipe(gulp.dest("./"));
-	gulp.src(fileSets).pipe(concat("kaphy.min.js")).pipe(uglify()).pipe(gulp.dest("./"));
+	gulp.src("./js/**/*.js").pipe(concat("kaphy.js")).pipe(gulp.dest("./"));
+	gulp.src("./js/**/*.js").pipe(concat("kaphy.min.js")).pipe(uglify()).pipe(gulp.dest("./"));
 	return;
 });
