--- js/canvas/init.js
@@ -1,9 +1,0 @@
-if(!KAPhy.Canvas) {
-  KAPhy.Canvas = {
-    configured: false,
-    relWidth: null,
-    relHeight: null,
-    element: null,
-    context: null
-  };
-}
