+++ js/physics/entities/line.js
@@ -1,21 +1,19 @@
-if (!KAPhy.Physics.Line) {
-  KAPhy.Physics.Line = function(config) {
-    this.one = config.one || new KAPhy.Physics.Vector2();
-    this.two = config.two || new KAPhy.Physics.Vector2();
-    
-    this.one = new KAPhy.Physics.Vector2(KAPhy.Canvas.toPixels(this.one.x), KAPhy.Canvas.toPixels(this.one.y));
-    this.two = new KAPhy.Physics.Vector2(KAPhy.Canvas.toPixels(this.two.x), KAPhy.Canvas.toPixels(this.two.y));
+KAPhy.Physics.Line = function(config) {
+ this.one = config.one || new KAPhy.Physics.Vector2();
+ this.two = config.two || new KAPhy.Physics.Vector2();
 
-    /** Bounce Coefficient **/
-    this.bcf = config.bcf || 0.75;
+ this.one = new KAPhy.Physics.Vector2(KAPhy.Canvas.toPixels(this.one.x), KAPhy.Canvas.toPixels(this.one.y));
+ this.two = new KAPhy.Physics.Vector2(KAPhy.Canvas.toPixels(this.two.x), KAPhy.Canvas.toPixels(this.two.y));
 
-    /** Radius **/
-    this.rad = config.rad || 4;
-  };
-  KAPhy.Physics.Line.prototype.draw = function() {
-    var sw = KAPhy.Draw.getStrokeWeight();
-    KAPhy.Draw.strokeWeight(this.rad * 2);
-    KAPhy.Draw.line(this.one.x, this.one.y, this.two.x, this.two.y);
-    KAPhy.Draw.strokeWeight(sw);
-  };
-}
+ /** Bounce Coefficient **/
+ this.bcf = config.bcf || 0.75;
+
+ /** Radius **/
+ this.rad = config.rad || 4;
+};
+KAPhy.Physics.Line.prototype.draw = function() {
+ var sw = KAPhy.Draw.getStrokeWeight();
+ KAPhy.Draw.strokeWeight(this.rad * 2);
+ KAPhy.Draw.line(this.one.x, this.one.y, this.two.x, this.two.y);
+ KAPhy.Draw.strokeWeight(sw);
+};
