--- js/ui/init.js
@@ -1,5 +1,0 @@
-if(!KAPhy.Interface) {
-  KAPhy.Interface = {
-    selected: null
-  };
-}
